refactor router handling out to separate file
